Fixing issue #328.  Changing to non-owning pointer.
